Fix event.next after EVENTS_LOST
When recording the id of the latest dropped event, we need to sort the list of
ids as integers. Unfortunately in the forward port of the events.from patch from
clearwater-lcm-sp1 I accidentally sorted the list of ids as strings.

- : bool = false

- : bool = true

This corrupts the stored id, resulting in multiple EVENTS_LOST where only one
is expected.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>