Disable test case that uses API call
The mock API feature is not quite feature-full to fully mock API calls as this
test case attempts to make use of. There is a Helpers.call_api_functions
(indirectly) in the codepath which fails since the mock database created does
not persist as far as the server that handles the API call.

We'll disable this test for now which is error-ing (not failing) until we have
time to properly provide a mock API.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>