squeezed: if running on xen, discover domain 0 memory via the balloon driver
We must ensure that xapi and squeezed agree on the amount of memory to
'assign' to dom0, so that the host free memory calculations are correct.
Ideally xapi should delegate all of this to squeezed.

For the moment, both xapi and squeezed will query the balloon driver first
since this number is most accurate, and fall back to /proc/meminfo (which
is usually a smaller number than expected).

Signed-off-by: David Scott <dave.scott@citrix.com>
Signed-off-by: Rob Hoes <rob.hoes@citrix.com>