CA-137468: Wait for ofport to settle before modifying ovs rules
The database may take some time for this value to become valid. Until then it
may return -1. In the VIF plug case we wait for this value to become valid
using the wait-until command provided with ovs-vsctl. For the VIF unplug case
we do not wait because once the value is -1 it will never become not -1 again.
In the case where the ofport returned is -1 (now only the unplug case) we no
longer delete the rules in the flow table since, in some versions of ovs,
the -1 can act as a wildcard in the deletion, deleting far more than desired.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>