CA-140545: wrapper should exit after launching vncterm
To avoid an instance of fe hanging around for each vncterm we launch,
fork vncterm and exit rather than calling exec.

This means that the script has to write the child pid ($!) rather than
its own pid ($$) to xenstore.

This commit also removes the exec of xenstore-write in the xiu branch,
since it should return straight away.

Signed-off-by: John Else <john.else@citrix.com>