CP-8987: Use Sexp for Xenops header serialisation
We want to be able to extend the Xenops record in the suspend image in future
versions. We therefore require a serialisation format that supports optional
parsing. Sexplib has such functionality with sexp_option allowing us to add new
fields to new versions and still parse older versions with these defaulting to
None.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>