CP-9093: decrease the inhibit period of fs_usage alert
It seems that the inhibit period of fs_usage alert is set too long (one week).
The consequence is that the alerts will be sent out much less frequently
(not showing much emergency as we want). Also it will mute the alarm for a week
long even if a different occurence shows up later on (after the first occurence
being addressed already) during this period of time.

Signed-off-by: Zheng Li <dev@zheng.li>