CP-8987: Deserialise the Xenops record on restore
We want to make sure we're receiving a valid suspend image so, as well as
reading in the right amount of length for the record (checking that the
macro-structure of the suspend image is correct), we should also attempt to
deserialise the Xenops_record to check that it is valid on a deeper level. For
now, we are not doing anything with it but the deserialisation alone checks the
validity.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>