Switch from domid- to uuid-based private XS keys in classic backend
The "private" xenstore keys are keys that are used to store
xapi/xenopsd-specific information for use by device hotplug scripts and
similar. The classic backend used to store these at /xapi/<domid>/private/.
This patch moves them to /xapi/<uuid>/private. This makes the classic backend
more similar to the libxl backend, such that the hotplug scripts can be merged.

We have to be careful in particular with localhost migration (and anything
using localhost migration, such as VDI migration), where multiple domains exist
for the same VM at the same time on the same host. To avoid race conditions, we
must keep the hotplug keys in domid-based subtrees (moved from
/xapi/<domid>/hotplug to /xapi/<uuid>/hotplug/<domid>). We also must be sure to
not accidentally remove the keys we still need on domain destroy.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>