CA-140402: set max-idle to 5000ms on OVS
Max-idle is an OVS setting that determines the idle timeout of flows in the
kernel.  Upstream OVS had reduced the timeout from 5000ms to 1500ms, causing
kernel flows to be removed sooner. For flows that send packets with an interval
that is a little larger than 1500ms, this means that every packet will result
in an upcall to the OVS userspace. Tests have shown a relatively large impact
on the dom0 CPU usage due to the reduced timeout.

This patch puts the max-idle timeout back to 5000ms. Do this in xcp-networkd
when it starts up.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>