Ensure that the Device_common.get_private_path does not raise an exception
Also add some extra logging in Xenstore_watch.

Returning a random string on error is not very neat, but we must avoid exceptions here.
This patch must be followed soon by a patch that changes the callers of
get_private_path{_of_device} to call by UUID, so that this code can go away.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>