xenlight: use an asynchronous way to report the outcome of asynchronous libxl calls
Sometimes an asynchronous libxl call may return and call back before the
initiating function returns. When this happens, the callback occurs on the same
thread as the initiating function and must not block.

We used to use the synchronous Event module to pass the result from the
callback back to the initiator, which does not work in this case. We replace it
by a custom event module with blocking reads and non-blocking writes.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>