CA-141502: Making message_destroy more robust
While destroying a message with a given ref, if
the ref does not exit under refs dir, a handle_invalid
error is thrown and the message is left undestroyed.

This patch reads the refs from all the messages under
/var/xapi/blobs/messages and matches it with the given
ref. Once the match is found, it retrives that message
file and is then destroyed.
Signed-off-by: Ravi Pandey <ravi.pandey@citrix.com>