CP-9366: Require HVM feature-flags (or PV drivers) strictly.
When attempting a lifecycle operation on an HVM guest, we now
require any relevant feature-flag (or PV drivers) in the same way
as when calculating the allowed_operations list.

This undoes another part of the behaviour-change (but not
code-change) of my earlier commit
43592b1185789a96ef9728ab761a0940639cc7cf

This returns behaviour for HVM Windows without guest-agent
(without PV drivers) to the same as before any CP-9366 changes.

In a future change we might make this configurable.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>