networkd: add Cisco UCS enic workaround check
When on the Linux bridge backend and when the "enic" driver is loaded, we need
to apply a workaround (implemented in the following commit).

The enic driver is for Cisco UCS devices. The current driver adds VLAN0 headers
to all incoming packets, which confuses certain guests OSes.  The workaround
constitutes added a VLAN0 Linux device to strip those headers again.

The patch moves the "backend_kind" ref outside the Bridge module, and adds an
"need_enic_workaround" function.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>