xe host-apply-edition: avoid race in error reporting
When host.apply_edition raises a license_checkout_error exception, we look at
the recent alerts (messages) to find out the reason for the failure. We seem to
currently miss such an alert every now and then, possibly due to rounding
errors in the used timestamp.  Make this less likely by subtracting a second.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>