Maintain a mapping of SMAPIv2 SR references to xapi-storage SRs
In the SMAPIv2 the client generates the SR name (!) and tells this to
the implementation. In the xapi-storage world, each SR can name itself
how it likes. To make these worlds meet, the daemon needs to maintain
a mapping over SR.attach/SR.detach.

The current implementation simply uses an in-memory hashtable. This
should be persisted to somewhere which is reset on host reboot.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>