CA-148772: support consoles served over Unix domain sockets
Xenopsd with libxl + upstream qemu prefers to bind consoles to
named sockets like

  /var/run/non-persistent/xenospd/consoles/VM/<uuid>

rather than anonymous localhost port numbers. Naming consoles is
a good idea because it prevents accidentally connecting to the
wrong port number.

This patch allows Xapi to connect to these Unix domain socket
consoles. Note there is still a lingering problem with upstream
qemu and XenCenter's built-in RFB client which causes the
keyboard not to work.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>