CAR-483: relax cross-pool migration power states check
In theory we should be able to support migration with any power states, but we
leave the `Paused state out for the moment. There is some underlying
implementation issue associated with the `Paused state (even for the normal
intra pool case). Moreover it is not a user visible power state anyway.

Signed-off-by: Zheng Li <dev@zheng.li>