CAR-483: add support for cross/intra-pool migrating a suspended VM
This mainly involves copying the suspend_VDI along the process.

At the API level, the caller can specifiy the destination SR for the suspend
VDI just like for any other VDIs. If unprovided,  a remote SR will be chosen
automatically based on the suspend_image_SR settings of the remote pool/host.

Signed-off-by: Zheng Li <dev@zheng.li>