Database: add a 'type' to each column
The types are:
 - String: most fields
 - Set: all Set(_)
 - Pairs: all Map(_,_)

This allows a database implementation to avoid serialising sets and maps
to strings all the time.

This patch also adds 'with sexp' to the schema so it can be easily read
and written. Eventually we will be able to separate the datamodel and
the schema and have the database load the schema at runtime.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>