Database: use the same 'remove' function signature everywhere
This allows us to put the 'remove' function into the MAP signature
for consistency. Now Row, Table, TableSet all have a remove which
takes a Time.t which allows tracking of delete events. We only
use these in the Table to track deleted Rows, but maybe in future
we could track deleted Tables or fields later?

Signed-off-by: David Scott <dave.scott@eu.citrix.com>