CA-141580: Fix dom0 console handler
Xenopsd no longer manages dom0 and so Xapi cannot ask it for the address of
dom0's console. Instead it should use the information it has in the database.

Incidentally, this information was no longer correct. Dom0 has had its console
on 5901 for a while it seems but the database used to get updated correctly
when Xenopsd managed dom0.

There is a patch in dom0 to change this back to 5900 to match Xapi
expectations.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>