CA-148986: Handle cancellation of cross-host VDI.copy
If Sparse_dd_wrapper.dd completes successfully, we wait for the remote
import task to no longer be in `pending state, and then handle the
resulting status of the task. If the task's status isn't `success then
something has gone wrong on the receiving end.

If Sparse_dd_wrapper.dd doesn't complete successfully, we know that
sparse_dd has been killed and we assume that the receiving xapi will
try to clean up after getting an EOF. In this case we wait for the
remote import task to no longer be in `pending state, and then re-raise
the original exception.

Signed-off-by: John Else <john.else@citrix.com>