CAR-483: Ask Importexport.remote_metadata_export_import to return the generated VMs
As we added the copy mode to Importexport.remote_metadata_export_import, the
generated VMs will not preserve the same uuids as the imported VMs with this
mode turned on. The only chance we can only get hold of the references of the
generated VMs is to ask Importexport.remote_metadata_export_import return them.

Signed-off-by: Zheng Li <dev@zheng.li>