CA-157775: Send content-length when exporting metadata
Instead of having the tar library stream the metadata directly to the remote
location, we now stream it through a pipe into a string. This lets us
read length of the tar data, and use this as the content-length supplied
in the HTTP response.

This allows the sender to use a cached stunnel; without sending the
content-length the receiver would wait forever as it couldn't determine
that all data had been sent.

We now also respond with HTTP 500 if anything goes wrong with the
metadata export; previously we would just write the tar footer and
continue as normal.

Signed-off-by: John Else <john.else@citrix.com>
Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>