CAR-483: return the ref of the new VM after cross-pool migration
This is useful for the normal SXM, as the caller can immediately get a handler
to the new VM in the destination pool (instead of having to find it out based
on the persistent uuid). This is more so for the copy mode, as the new VM will
not share the same uuid as the original one.

Signed-off-by: Zheng Li <dev@zheng.li>