CAR-483: workaround to let HA cope with SXM
This is to fix the problem described in CP-9049, where the migrated VM was
started by HA before the whole migration process completed.

The root problem is that we don't have a staging mechanism for VM.import, so
that autonomous components like xha will act upon a VM immmediately after it
being imported, and leave us with no chance to do any further tweaks on it if
needed.

This workaround involves:

 * First to delay the next HA plan cycle a little bit to make sure it won't
   happen during, or immediately after, the very short period of VM metadata
   import.
 * Disable the HA runtime flag on the VM immediately after the import, so that
   it can escape from the HA for the them being.
 * Restore the HA flag after we've done all the tweaks so that it now starts to
   be covered by HA.

Signed-off-by: Zheng Li <dev@zheng.li>