Refactor handling of IGD_passthru
Instead of searching for integrated GPUs which need to be passed through
when generating the xenopsd metadata, we now do it as part of
Vgpuops.add_pcis_to_vm and add a flag to the platform map if any of the
devices to be passed through are integrated GPUs.

Xapi_xenops.builder_of_vm now uses this flag to integrate it should send
the IGD_passthrough video mode to xenopsd.

This removes the Helpers.get_localhost call from
Xapi_xenops.builder_of_vm, which was breaking the unit tests.

Signed-off-by: John Else <john.else@citrix.com>