Integrate Xapi_pgpu.update_gpus with new host and PGPU fields
For the system display device, update_gpus needs to take into account
whether the host display is enabled or not. If the host display is
enabled, we can't support any VGPU types on the system display device.

We also need to determine whether the system display device is
accessible to dom0. For new devices we haven't seen before, or if the
host is booting, we determine this from the boot config. In all other
cases, we use the value of PGPU.dom0_access already in the database.

Signed-off-by: John Else <john.else@citrix.com>