CP-11136: Handle Set(Int) in "exposed_type" functions.
First part of a fix.

The work for CP-11136 in xen-api added a field of type Set(Int)
to the Host, to hold the set of Virtual Hardware Platform
Versions that it offers. The "exposed_type" functions here in
the SDK generation code did not handle Set(Int) but threw an
assertion error.  Now they return "long[]".

Still to do: handle Set(Int) in the functions
simple_convert_from_proxy and convert_to_proxy in
csharp/gen_csharp_binding.ml

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>