CA-162159: bump the fd limit in xenconsoled to 16384
init.d/xenservices is limiting the max number of fds in xenconsoled to 4096.
xenconsoled uses 4 fds for each vm, so this is limiting the number of vms in
the host to 1000.

Increasing the max number of fds in xenconsoled to 16384 will let the host
start up to 4000 vms, which is a more reasonable limit given the other
current limits in dom0.