CP-11511: allow storage backends to choose a datapath
The storage backends should choose the default datapath for a given
VDI, which could be: tapdisk3, blkback or even qdisk.

This patch extends the existing 'backend-kind' mechanism to allow an
SMAPIv1/2 backend to return a choice via the 'xenstore_data'. This is
a port of the existing libxl mechanism in c/s 8fb7bcb4

Signed-off-by: David Scott <dave.scott@eu.citrix.com>