CA-164806: Upgrade allow-gpu-passthrough recommendations
The patch adds an upgrade rule that modifies the VM.recommendations field when
upgrading from pre-Cream to Cream. For all VMs that have allow-gpu-passthrough
set to 0, the rule sets allow-gpu-passthrough to 1 and additionally adds
allow-vgpu and sets this to 0. The assumption is that the only pre-Cream VMs
that have allow-gpu-passthrough = 0 are HVM Linux VMs.

The reason for doing this is that HVM Linux guests gained the ability to use
GPU passthrough in Cream, but this won't be automatically reflected in the
VM.recommendations (used by e.g. XenCenter to determine whether the
functionality is allowed). For the future, we should find a better way of
upgrading this field.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>