CA-165131: fix the backend-kind override mechanism
The SM backend can return a backend-kind in the attach_info:xenstore_data.
If the user knows what they are doing they can override this with
VBD.other_config. This user override mechanism is currently used by
the RawHBA SR. This patch fixes it by removing a duplicate definition of
"kind".

Signed-off-by: David Scott <dave.scott@eu.citrix.com>