Refactor Rrd_reader and Rrd_writer
Instead of keeping a file descriptor as the state of a FileReader and
mmap'ing it each tick, mmap once and keep the resulting Cstruct as the
state.

This requires an interface change. A number of pages to share must now
be specified when creating a FileWriter; this means that programs
writing RRDs must know ahead of time the maximum amount of data they
will need to share. This means FileWriter now works more like
PageWriter.

Signed-off-by: John Else <john.else@citrix.com>