CAR-483: align the mode of mirroring DP with that of the original VBD
Though we have no intention of "write", here we set the mode of mirroring DP
the same as that of the associated VBD (which is, always RW, for mirroring
case). This avoids problem when we need to start/stop the VM along the migration.

The issue was originally reported as CA-165664.

Signed-off-by: Zheng Li <dev@zheng.li>