CAR-483: introduce presentative_datapath_of_vbd
We now support storage migration of stopped and suspended VM, whose VBDs are
not plugged and there are no associated DPs either. However, as we allow VM
power status change (e.g. VM start/stop) during the migration process, these
VBD might be plugged and DP might be created right in the middle of migration.

For that reason, we need a presentative path for the potential DP disregarding
the power status of the VM and even existence of the underlying VBD (e.g. for
the suspend_VDI).

Signed-off-by: Zheng Li <dev@zheng.li>