CAR-483: remove VM power status restricitons for VDI.pool_migrate.
... also previously we only allow the migration when the associated VM can see
both the source and the destination SRs at the same time, and this is not
strictly necessary now. Suppose we want to migrate a VDI from the local SR of host
A to local SR of host B, this can be done given that the associated VM will be
automatically migrated to host B from host A.

The only restriction is that after the "potential" migration, the associated VM
must remain in a valid status. E.g., if after the migration (with no other VDI
changed), the associated VM has two VDIs in two different local SRs (of two
different hosts), then the VM is no longer in a valid status as it will never be
able to start on any host. Cases like this will be rejected from day one.

Signed-off-by: Zheng Li <dev@zheng.li>