Use buffered I/O for the redo-log
We only need the redo-log to persist until the next reboot, so we
don't have to force it to hit the disk each time. Using the page
cache is safe (and so would running it on a ramdisk or tmpfs)

This increases performance up to about 4k ping-pong RPCs/sec. Without
persistence it manages about 6k ping-pong RPCs/sec.

Signed-off-by: David Scott <dave.scott@citrix.com>