Revert "Handle invalid edition on xapi startup"
This reverts commit c01a63fa92af3df02fa1ba79be404bbb9be89e1d.

This patch of code first checks the licenses available and then defaults
to minimum edition if the existing license is not among the set of
possible licenses. This results in issues like CA-167469.

We do not need this as we want xapi to tell v6d just the license name
and it would v6d's responsibilty to decide what to do with that.