CA-170667: Start SMAPIv1 proxies before dbsync on slave
Dbsync on a slave needs to be able to do a storage call for PBDs that are
already attached (e.g. when xapi restarts during a master switchover).  The
storage calls need the SMAPIv1 proxies to be running. The SMAPIv1 proxies need
the DB to be up, and therefore need a master connection on a slave.  Therefore,
on a slave, the proxies need to be started after establishing the master
connection, but before the dbsync.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>