CP-12511 xapissl: TLSv1.2 and back-compat option
Change the xapissl script to use TLSv1.2 only by default, and to
accept an option to tell it to allow SSLv3 (and associated ciphers)
so as to allow connections from XenServer 6.5 and earlier (for
rolling pool upgrade or cross-pool migration).

First of a series of related commits. This one on its own will
break intra-pool communication between XenServers that use
stunnel 4.15 rather than something newer; this will change in a
subsequent commit.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>