CP-12511: When starting stunnel, allow SSL if RPU in progress
This is for backward compatibility during rolling pool upgrade,
allowing incoming connections from older XenServer hosts (e.g. v6.5)
that are not yet upgraded to use modern TLS for their outgoing
connections.

On RPU completion we do not disable the backward compatibility mode.
That might be implemented in future work.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>