CP-12511 Temp commit: use SSL back-compat mode
Temporary code-change to always allow old-style SSLv3 incoming
connections as in XenServer 6.5 and earlier.

This is needed for the time being even with latest development versions
of XenServer.  Once XenServer has a newer stunnel that can use TLSv1.2
(protocol and ciphersuites) for outgoing connections, we can undo this
commit without breaking inter-host communications.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>