CP-12633 ssl_legacy controls listening stunnel
Link up the Host.ssl_legacy field and its setter to control
stunnel.  When needed, the setter rewrites the stunnel config file
and restarts stunnel (for incoming connections).

When we generate the stunnel config, we consult ssl_legacy instead
of checking whether a rolling pool upgrade is in progress.

All of this is for incoming connections only at this stage.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>