CA-172039: Shutdown VM before reverting its fields
The Message_forwarding layer reverted the some of the VM's fields on the master
before sending the operation to relevant host. This will generate some Xapi
events and can cause issues if they result in updates to Xenopsd metadata.

As a concrete example, consider a halted VM that has no GPU passthrough; then:

  1. Snapshot the VM;
  2. Assign it a GPU for passthrough;
  3. Start the VM;
  4. Revert the VM to the snapshot (1).

This will cause an error because the PCI that is assigned for passthrough is
stored in the VM.other_config field. When this is reverted it will generate
a Xapi event which results in Xenopsd being told that the VM has no PCIs
anymore and so Xenopsd removes its knowledge of the PCI device.

However, this is racing with the VM.hard_shutdown that is part of the VM.revert
call which can then fail when it gets to the PCI.unplug operation since the PCI
has been removed from the Xenopsd database.

This changeset moves the VM.hard_shutdown logic before the reversion of the VM
record.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>