CA-173622: Enable session.logout()
Several API examples exist, including one in xen-api, that use
session.logout directly without going through session.xenapi or
session.xenapi.session.  While this is not the correct way to
logout of the XAPI session, logout is one of the special cases
that needs to be supported from the session object.

Signed-Off-By: Bob Ball <bob.ball@citrix.com>