CAR-1744: Add a fail_immediately policy for message_forwarding's retry logic
In some cases we'd like to fail immediately (and to notifiy the caller/user) if
other operation is in progress.

Instead of duplicating existing code in higher level functions such as
with_vm_operation for this special case (no retry), we simply add this special
policy and expose the policy parameter to the caller.

Note that even though it's called fail_immediately, we deliberately keep a
minimal interval coverage of 0-3 seconds. Two benefits:

 * This could avoid staggering by accident
 * This could help to avoid DDOS

Signed-off-by: Zheng Li <dev@zheng.li>