CA-175353: resync networking after VMs have been resynced
When performing a bond/VLAN resync we need the ability to call the
xenopsd VIF.move API to move the backend device onto the right
bridge or switch. Previously these attempts would block because
the events from xenopsd thread wasn't running. To start the
events from xenopsd thread we need to have resynchronised the VM
state, which also requires the storage to be online.

Note this makes the case where a resync is done after a `service
xapi restart` on a live system, with plugged storage and running
VMs, more similar to the case where a resync is done after a
reboot.

Signed-off-by: David Scott <dave.scott@citrix.com>