Preliminary support for clone-on-boot / reset-on-boot
Upon VDI.epoch_begin with persistent = false, we call `Volume.stat`
to discover the possible URIs. We look first for a datapath plugin
which can perform clone-on-boot natively (e.g. by truncating vhd
leaves) and delegate to that. If a datapath plugin doesn't exist
then we will fall back to the Volume.clone / Volume.destroy API. This
will be done in a later patch.

We expose the VDI_RESET_ON_BOOT/2 capability if the volume plugin
supports VDI_CLONE. This isn't ideal as it won't cover the case where
a LUN-per-VDI SR which doesn't support VDI.CLONE nevertheless can
support clone-on-boot if the tapdisk datapath is used.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>